+++ drivers/ec_host_cmd_periph/ec_host_cmd_simulator.c
@@ -9,8 +9,7 @@
 #include <errno.h>
 
 #include <zephyr/device.h>
-#include <zephyr/drivers/ec_host_cmd_periph.h>
-#include <zephyr/kernel.h>
+#include <zephyr/drivers/ec_host_cmd_periph/ec_host_cmd_periph.h>
 #include <string.h>
 
 #ifndef CONFIG_ARCH_POSIX
