+++ subsys/mgmt/ec_host_cmd/ec_host_cmd_handler.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <zephyr/drivers/ec_host_cmd_periph.h>
+#include <zephyr/drivers/ec_host_cmd_periph/ec_host_cmd_periph.h>
 #include <zephyr/mgmt/ec_host_cmd.h>
 #include <zephyr/devicetree.h>
 #include <zephyr/kernel.h>
